+++ africanus/rime/tests/test_predict.py
@@ -35,7 +35,7 @@ def rc(*a, **kw):
 ])
 def test_predict_vis(corr_shape, idm, einsum_sig1, einsum_sig2,
                      a1j, blj, a2j, g1j, bvis, g2j):
-    from africanus.rime.predict2 import predict_vis
+    from africanus.rime.predict import predict_vis
 
     s = 2       # sources
     t = 4       # times
@@ -112,8 +112,8 @@ def test_dask_predict_vis(corr_shape, idm, einsum_sig1, einsum_sig2,
 
     da = pytest.importorskip('dask.array')
     import numpy as np
-    from africanus.rime.predict2 import predict_vis as np_predict_vis
-    from africanus.rime.dask import predict_vis2
+    from africanus.rime.predict import predict_vis as np_predict_vis
+    from africanus.rime.dask import predict_vis
 
     # chunk sizes
     sc = (2, 3, 4)    # sources
@@ -162,13 +162,13 @@ def test_dask_predict_vis(corr_shape, idm, einsum_sig1, einsum_sig2,
     da_base_vis = da.from_array(base_vis, chunks=(rrc, cc) + corr_shape)
     da_g2_jones = da.from_array(g2_jones, chunks=(tc, ac, cc) + corr_shape)
 
-    model_vis = predict_vis2(da_time_idx, da_ant1, da_ant2,
-                             da_a1_jones if a1j else None,
-                             da_bl_jones if blj else None,
-                             da_a2_jones if a2j else None,
-                             da_g1_jones if g1j else None,
-                             da_base_vis if bvis else None,
-                             da_g2_jones if g2j else None)
+    model_vis = predict_vis(da_time_idx, da_ant1, da_ant2,
+                            da_a1_jones if a1j else None,
+                            da_bl_jones if blj else None,
+                            da_a2_jones if a2j else None,
+                            da_g1_jones if g1j else None,
+                            da_base_vis if bvis else None,
+                            da_g2_jones if g2j else None)
 
     model_vis = model_vis.compute()
